refactor: unify search components for universities and study programs
