+++ frontend/src/components/Universities/Universities.tsx
@@ -2,10 +2,9 @@ import { useState, use } from "react";
 import { RootContext } from "../../contextData/RootContext";
 import { Helmet } from "react-helmet-async";
 import { GetAllUniversities } from "./GetAllUniversities";
-import { SearchUniversities } from "./SearchUniversities";
-import { SearchStudyPrograms } from "./SearchStudyPrograms";
+import { UnifiedSearch } from "./UnifiedSearch";
 
-const TABS = ["search", "findPrograms", "browseAll"];
+const TABS = ["search", "browseAll"];
 
 function Universities() {
   const { t } = use(RootContext);
@@ -62,8 +61,7 @@ function Universities() {
 
         <div className="w-full">
           {activeTab === "browseAll" && <GetAllUniversities />}
-          {activeTab === "search" && <SearchUniversities />}
-          {activeTab === "findPrograms" && <SearchStudyPrograms />}
+          {activeTab === "search" && <UnifiedSearch />}
         </div>
       </div>
     </>
