+++ frontend/src/locales/en.json
@@ -336,11 +336,12 @@
     "title": "Find programs and universities",
     "browseAll": "Browse All",
     "search": "Search",
-    "findPrograms": "Find Study Programs",
-    "searchPlaceholder": "Search by name, city, or acronym…",
-    "studyProgramsPlaceholder": "Search by study program name…",
+    "searchAllPlaceholder": "Search universities and study programs…",
+    "universitiesSection": "Universities",
+    "studyProgramsSection": "Study programs",
     "noResults": "No universities found.",
     "noStudyProgramResults": "No study programs found.",
+    "noResultsAtAll": "No universities or study programs found.",
     "city": "City",
     "entity": "Entity",
     "entities": {
