+++ frontend/src/locales/sr.json
@@ -336,11 +336,12 @@
     "title": "Nađite programe i univerzitete",
     "browseAll": "Pregledaj sve",
     "search": "Pretraži",
-    "findPrograms": "Pronađi studijske programe",
-    "searchPlaceholder": "Pretraži po nazivu, gradu ili akronimu…",
-    "studyProgramsPlaceholder": "Pretraži po nazivu studijskog programa…",
+    "searchAllPlaceholder": "Pretraži univerzitete i studijske programe…",
+    "universitiesSection": "Univerziteti",
+    "studyProgramsSection": "Studijski programi",
     "noResults": "Nisu pronađeni univerziteti.",
     "noStudyProgramResults": "Nisu pronađeni studijski programi.",
+    "noResultsAtAll": "Nisu pronađeni ni univerziteti ni studijski programi.",
     "city": "Grad",
     "entity": "Entitet",
     "entities": {
